move function doStepJump from rollout_canary.go to rollout_status.go
