+++ api/v1beta1/rollout_types.go
@@ -459,8 +459,8 @@ type BlueGreenStatus struct {
 	UpdatedReadyReplicas int32 `json:"updatedReadyReplicas"`
 }
 
-// GetSubStatus returns the ethier canary or bluegreen status
-// Deprecated: use controller GetUnifiedStatus
+// GetSubStatus returns the either canary or blueGreen status
+// TODO: It will be gradually deprecated in the future, moving towards using UnifiedStatus to read status uniformly
 func (r *RolloutStatus) GetSubStatus() *CommonStatus {
 	if r.CanaryStatus == nil && r.BlueGreenStatus == nil {
 		return nil
