+++ pkg/controller/rollout/rollout_canary.go
@@ -312,41 +312,6 @@ func (m *canaryReleaseManager) doCanaryPaused(c *RolloutContext) (bool, error) {
 	return false, nil
 }
 
-// doStepJump implements the common logic for both canary and bluegreen rollout strategies
-// to handle step jumping based on NextStepIndex.
-func doStepJump(rollout *v1beta1.Rollout, newStatus *v1beta1.RolloutStatus, steps []v1beta1.CanaryStep, workloadReplicas int) (jumped bool) {
-	status := GetUnifiedStatus(newStatus)
-	if status.IsNil() {
-		klog.InfoS("doStepJump skipped: unified status is nil", "rollout", klog.KObj(rollout))
-		return false
-	}
-	klog.InfoS("will do step jump", "steps", len(steps), "updatedReplicas", *status.UpdatedReplicas,
-		"nextStepIndex", status.NextStepIndex, "rollout", klog.KObj(rollout))
-	if nextIndex := status.NextStepIndex; nextIndex != util.NextBatchIndex(rollout, status.CurrentStepIndex) &&
-		nextIndex > 0 && nextIndex <= int32(len(steps)) {
-		currentIndexBackup := status.CurrentStepIndex
-		currentStepStateBackup := status.CurrentStepState
-		// update the current and next stepIndex
-		status.CurrentStepIndex = nextIndex
-		status.NextStepIndex = util.NextBatchIndex(rollout, nextIndex)
-		nextStep := steps[nextIndex-1]
-		// compare next step and current step to decide the state we should go
-		nextStepReplicas, _ := intstr.GetScaledValueFromIntOrPercent(nextStep.Replicas, workloadReplicas, true)
-		if int32(nextStepReplicas) == *status.UpdatedReplicas {
-			status.CurrentStepState = v1beta1.CanaryStepStateTrafficRouting
-		} else {
-			status.CurrentStepState = v1beta1.CanaryStepStateInit
-		}
-		status.LastUpdateTime = &metav1.Time{Time: time.Now()}
-		klog.InfoS("step jumped", "rollout", klog.KObj(rollout),
-			"oldCurrentIndex", currentIndexBackup, "newCurrentIndex", status.CurrentStepIndex,
-			"oldCurrentStepState", currentStepStateBackup, "newCurrentStepState", status.CurrentStepState)
-		return true
-	}
-	klog.InfoS("step not jumped", "rollout", klog.KObj(rollout))
-	return false
-}
-
 func (m *canaryReleaseManager) doCanaryJump(c *RolloutContext) (jumped bool) {
 	if c.Workload == nil {
 		return false
